requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
cinemagoer==2023.5.1
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.4
daphne==4.2.0
dj-static==0.0.6
Django==5.2.3
django-appconf==1.1.0
django-cors-headers==4.7.0
django-user-agents==0.4.0
django-video-encoding==1.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
googletrans==4.0.2
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==6.0.0
numpy==2.3.1
opencv-python==4.11.0.86
pandas==2.3.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2025.2
qrcode==8.2
requests==2.32.4
scipy==1.16.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
static-ranges==0.2.0
static3==0.7.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
webvtt-py==0.5.1
zope.interface==7.2
